Changed code for top button bar
I need 6 buttons so addes space for two.  Old code was all wrong and misused the button group feature of Bootstrap so redid it as a toolbar plus one more button far right.  This changed header_plus and index and I will have to change all the other screens to match as well.