Issue 32. Improve method of doing calculation
Participants hated putting in tags so now I converted a button in the text editor to be a calculator button.  Select the text to be evaluated and click Calculate.  The text will appear in italics (because I use the italics html tag to mark text to calculate.
Also replaced the old editor with niceEdit.  Had a bug with the old one as it kept adding paragraph tags to bold and italics.